+++ web/src/main/java/com/infoshareacademy/exception/RecipeUploadedFileNotFound.java
@@ -1,8 +1,10 @@
 package com.infoshareacademy.exception;
 
 public class RecipeUploadedFileNotFound extends Exception {
+
     public RecipeUploadedFileNotFound() {
     }
+
     public RecipeUploadedFileNotFound(String message){
         super(message);
     }
