+++ web/src/main/java/com/infoshareacademy/servlet/AdminViewServlet.java
@@ -12,13 +12,10 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 @WebServlet("/admin-view")
 public class AdminViewServlet extends HttpServlet {
 
-    Logger logger = LoggerFactory.getLogger(AdminViewServlet.class);
     @Inject
     private TemplateProvider templateProvider;
